requirements.txt
streamlit
streamlit_code_editor
lxml
verovio